V0.8.0.3 Fix a bug that score will overflow when it large then int.MaxValue.

+++ C#/src/Hubble.Data/Hubble.Core/Index/InvertedIndex.cs
@@ -289,7 +289,7 @@ namespace Hubble.Core.Index
                             score = (int)rank;
                         }
 
-                        rank = fieldRank * wordRank * score;
+                        rank = (long)fieldRank * (long)wordRank * (long)score;
 
                         Query.DocumentResult docResult;
